+++ src/components.d.ts
@@ -20,6 +20,7 @@ import { InterfaceSbbChipAttributes } from "./components/sbb-chip/sbb-chip.custo
 import { InputUpdateEvent } from "./components/sbb-datepicker/sbb-datepicker.helper";
 import { InterfaceSbbDividerAttributes } from "./components/sbb-divider/sbb-divider.custom";
 import { InterfaceSbbExpansionPanelAttributes } from "./components/sbb-expansion-panel/sbb-expansion-panel.custom";
+import { InterfaceSbbFileSelectorAttributes } from "./components/sbb-file-selector/sbb-file-selector.custom";
 import { InterfaceFooterAttributes } from "./components/sbb-footer/sbb-footer.custom";
 import { InterfaceSbbFormFieldAttributes } from "./components/sbb-form-field/sbb-form-field.custom";
 import { InterfaceSbbHeaderActionAttributes } from "./components/sbb-header-action/sbb-header-action.custom";
@@ -65,6 +66,7 @@ export { InterfaceSbbChipAttributes } from "./components/sbb-chip/sbb-chip.custo
 export { InputUpdateEvent } from "./components/sbb-datepicker/sbb-datepicker.helper";
 export { InterfaceSbbDividerAttributes } from "./components/sbb-divider/sbb-divider.custom";
 export { InterfaceSbbExpansionPanelAttributes } from "./components/sbb-expansion-panel/sbb-expansion-panel.custom";
+export { InterfaceSbbFileSelectorAttributes } from "./components/sbb-file-selector/sbb-file-selector.custom";
 export { InterfaceFooterAttributes } from "./components/sbb-footer/sbb-footer.custom";
 export { InterfaceSbbFormFieldAttributes } from "./components/sbb-form-field/sbb-form-field.custom";
 export { InterfaceSbbHeaderActionAttributes } from "./components/sbb-header-action/sbb-header-action.custom";
@@ -605,6 +607,40 @@ export namespace Components {
           * The icon name we want to use, choose from the small icon variants from the ui-icons category from here https://icons.app.sbb.ch.
          */
         "iconName"?: string;
+    }
+    interface SbbFileSelector {
+        /**
+          * A comma-separated list of allowed unique file type specifiers.
+         */
+        "accept": string;
+        /**
+          * This will be forwarded as aria-label to the native input element.
+         */
+        "accessibilityLabel": string | undefined;
+        /**
+          * Whether the component is disabled.
+         */
+        "disabled": boolean;
+        /**
+          * Gets the currently selected files.
+         */
+        "getFiles": () => Promise<File[]>;
+        /**
+          * Whether more than one file can be selected.
+         */
+        "multiple": boolean;
+        /**
+          * Whether the newly added files should override the previously added ones.
+         */
+        "multipleMode": InterfaceSbbFileSelectorAttributes['multipleMode'];
+        /**
+          * The title displayed in `dropzone` variant.
+         */
+        "titleContent"?: string;
+        /**
+          * Whether the component has a dropzone area or not.
+         */
+        "variant": InterfaceSbbFileSelectorAttributes['variant'];
     }
     interface SbbFooter {
         /**
@@ -2018,6 +2054,10 @@ export interface SbbExpansionPanelHeaderCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLSbbExpansionPanelHeaderElement;
 }
+export interface SbbFileSelectorCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLSbbFileSelectorElement;
+}
 export interface SbbMenuCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLSbbMenuElement;
@@ -2244,6 +2284,12 @@ declare global {
     var HTMLSbbExpansionPanelHeaderElement: {
         prototype: HTMLSbbExpansionPanelHeaderElement;
         new (): HTMLSbbExpansionPanelHeaderElement;
+    };
+    interface HTMLSbbFileSelectorElement extends Components.SbbFileSelector, HTMLStencilElement {
+    }
+    var HTMLSbbFileSelectorElement: {
+        prototype: HTMLSbbFileSelectorElement;
+        new (): HTMLSbbFileSelectorElement;
     };
     interface HTMLSbbFooterElement extends Components.SbbFooter, HTMLStencilElement {
     }
@@ -2676,6 +2722,7 @@ declare global {
         "sbb-expansion-panel": HTMLSbbExpansionPanelElement;
         "sbb-expansion-panel-content": HTMLSbbExpansionPanelContentElement;
         "sbb-expansion-panel-header": HTMLSbbExpansionPanelHeaderElement;
+        "sbb-file-selector": HTMLSbbFileSelectorElement;
         "sbb-footer": HTMLSbbFooterElement;
         "sbb-form-error": HTMLSbbFormErrorElement;
         "sbb-form-field": HTMLSbbFormFieldElement;
@@ -3318,6 +3365,40 @@ declare namespace LocalJSX {
          */
         "iconName"?: string;
         "onToggle-expanded"?: (event: SbbExpansionPanelHeaderCustomEvent<any>) => void;
+    }
+    interface SbbFileSelector {
+        /**
+          * A comma-separated list of allowed unique file type specifiers.
+         */
+        "accept"?: string;
+        /**
+          * This will be forwarded as aria-label to the native input element.
+         */
+        "accessibilityLabel"?: string | undefined;
+        /**
+          * Whether the component is disabled.
+         */
+        "disabled"?: boolean;
+        /**
+          * Whether more than one file can be selected.
+         */
+        "multiple"?: boolean;
+        /**
+          * Whether the newly added files should override the previously added ones.
+         */
+        "multipleMode"?: InterfaceSbbFileSelectorAttributes['multipleMode'];
+        /**
+          * An event which is emitted each time the file list changes.
+         */
+        "onFile-changed"?: (event: SbbFileSelectorCustomEvent<File[]>) => void;
+        /**
+          * The title displayed in `dropzone` variant.
+         */
+        "titleContent"?: string;
+        /**
+          * Whether the component has a dropzone area or not.
+         */
+        "variant"?: InterfaceSbbFileSelectorAttributes['variant'];
     }
     interface SbbFooter {
         /**
@@ -4813,6 +4894,7 @@ declare namespace LocalJSX {
         "sbb-expansion-panel": SbbExpansionPanel;
         "sbb-expansion-panel-content": SbbExpansionPanelContent;
         "sbb-expansion-panel-header": SbbExpansionPanelHeader;
+        "sbb-file-selector": SbbFileSelector;
         "sbb-footer": SbbFooter;
         "sbb-form-error": SbbFormError;
         "sbb-form-field": SbbFormField;
@@ -4911,6 +4993,7 @@ declare module "@stencil/core" {
             "sbb-expansion-panel": LocalJSX.SbbExpansionPanel & JSXBase.HTMLAttributes<HTMLSbbExpansionPanelElement>;
             "sbb-expansion-panel-content": LocalJSX.SbbExpansionPanelContent & JSXBase.HTMLAttributes<HTMLSbbExpansionPanelContentElement>;
             "sbb-expansion-panel-header": LocalJSX.SbbExpansionPanelHeader & JSXBase.HTMLAttributes<HTMLSbbExpansionPanelHeaderElement>;
+            "sbb-file-selector": LocalJSX.SbbFileSelector & JSXBase.HTMLAttributes<HTMLSbbFileSelectorElement>;
             "sbb-footer": LocalJSX.SbbFooter & JSXBase.HTMLAttributes<HTMLSbbFooterElement>;
             "sbb-form-error": LocalJSX.SbbFormError & JSXBase.HTMLAttributes<HTMLSbbFormErrorElement>;
             "sbb-form-field": LocalJSX.SbbFormField & JSXBase.HTMLAttributes<HTMLSbbFormFieldElement>;